requirements.txt
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
requests==2.32.3
numpy==2.2.3
python-dotenv==1.0.1
geopy==2.4.1
